Change imports to remove circular deps & delete subrow
