+++ src/components/IndexProvider/index.ts
@@ -1,2 +1,1 @@
 export * from './IndexProvider';
-export * from './utilities';
