+++ src/components/IndexTable/components/Checkbox/Checkbox.tsx
@@ -10,12 +10,12 @@ import React, {
 
 import {useI18n} from '../../../../utilities/i18n';
 import {classNames} from '../../../../utilities/css';
-import {RowContext} from '../Row';
-import {useIndexValue} from '../../../IndexProvider';
+import {RowContext} from '../../../../utilities/index-table';
+import {useIndexValue} from '../../../../utilities/index-provider';
 import {Checkbox as PolarisCheckbox} from '../../../Checkbox';
 import {setRootProperty} from '../../../../utilities/set-root-property';
-
 import sharedStyles from '../../IndexTable.scss';
+
 import styles from './Checkbox.scss';
 
 export const Checkbox = memo(function Checkbox() {
