+++ src/components/IndexTable/components/Row/index.ts
@@ -1,3 +1,1 @@
 export * from './Row';
-export * from './context';
-export * from './hooks';
