--- src/components/IndexTable/components/SubRow/index.ts
@@ -1,1 +1,0 @@
-export * from './SubRow';
