+++ src/components/IndexTable/components/index.ts
@@ -1,5 +1,4 @@
 export * from './Cell';
 export * from './Row';
 export * from './Checkbox';
-export * from './SubRow';
 export * from './ScrollContainer';
