+++ src/components/index.ts
@@ -137,7 +137,6 @@ export {Image} from './Image';
 export type {ImageProps} from './Image';
 
 export {IndexProvider as UnstableIndexProvider} from './IndexProvider';
-export type {IndexProviderProps as UnstableIndexProviderProps} from './IndexProvider';
 
 export {IndexTable as UnstableIndexTable} from './IndexTable';
 export type {IndexTableProps as UnstableIndexTableProps} from './IndexTable';
