+++ src/utilities/index-provider/index.ts
@@ -1,2 +1,3 @@
+export * from './hooks';
 export * from './context';
 export * from './types';
