+++ src/utilities/index-table/index.ts
@@ -1,0 +1,2 @@
+export * from './hooks';
+export * from './context';
